@@ -2,8 +2,8 @@
   "name": "@scullyio/universal",
   "version": "0.0.1",
   "peerDependencies": {
-    "@angular/common": "^11.2.5",
-    "@angular/core": "^11.2.5"
+    "@angular/common": "^12.0.0",
+    "@angular/core": "^12.0.0"
   },
   "dependencies": {
     "tslib": "^2.0.0"
