@@ -2,5 +2,5 @@ export {Job} from './job'
 export {handleJobs} from './handleJobs'
 export {TaskWorker} from './TaskWorker'
 export {startWorkerListener, workerMessages$, addWorkerTask} from './startWorkerListener'
-export {getPool} from './taskPool'
+export {getPool, terminatePool} from './taskPool'
 export {Tasks} from './tasks.interface'
