@@ -50,60 +50,66 @@ export const startScully = async (url?: string) => {
         return routes.length;
       })
       .catch(() => 0);
-    Promise.all([numberOfRoutesProm, durationProm]).then(([numberOfRoutes, durations]) => resolve({ numberOfRoutes, durations }));
-  }).then(({ numberOfRoutes, durations }: { numberOfRoutes: number; durations: { [key: string]: number } }) => {
-    const duration = durations.Duration;
-    // tslint:disable-next-line:variable-name
-    const seconds = duration / 1000;
-    const singleTime = duration / numberOfRoutes;
-    const routesProSecond = Math.ceil((1000 / singleTime) * 100) / 100;
-    // console.table(durations)
-    stopProgress();
-    reloadAll();
-    log(`
+    Promise.all([numberOfRoutesProm, durationProm])
+      .then(([numberOfRoutes, durations]) => resolve({ numberOfRoutes, durations }));
+    /** stop progress spinner. */
+    numberOfRoutesProm.then(() => stopProgress());
+
+  })
+    .then(displayAndWriteStats)
+    .catch((e) => {
+      // stopProgress();
+      // console.error(e);
+      captureException(e);
+      process.exit(15);
+    });
+};
+
+function displayAndWriteStats({ numberOfRoutes, durations }: { numberOfRoutes: number; durations: { [key: string]: number; }; }) {
+  const duration = durations.Duration;
+  // tslint:disable-next-line:variable-name
+  const seconds = duration / 1000;
+  const singleTime = duration / numberOfRoutes;
+  const routesProSecond = Math.ceil((1000 / singleTime) * 100) / 100;
+  // console.table(durations)
+  stopProgress();
+  reloadAll();
+  log(`
 Generating took ${yellow(Math.floor(seconds * 100) / 100)} seconds for ${yellow(numberOfRoutes)} pages:
   That is ${yellow(routesProSecond)} pages per second,
   or ${yellow(Math.ceil(singleTime))} milliseconds for each page.
-  ${
-    durations.Traverse
+  ${durations.Traverse
       ? `
   Finding routes in the angular app took ${logSeconds(durations.Traverse)}`
-      : ''
-  }
+      : ''}
   Pulling in route-data took ${logSeconds(durations.Discovery)}
   Rendering the pages took ${logSeconds(durations.Render)}
 
-${
-  watch
-    ? `The server is available on "${yellow(`http${ssl ? 's' : ''}://${scullyConfig.hostName}:${scullyConfig.staticPort}/`)}"
+${watch
+      ? `The server is available on "${yellow(`http${ssl ? 's' : ''}://${scullyConfig.hostName}:${scullyConfig.staticPort}/`)}"
 ${yellow('------------------------------------------------------------')}
 Press ${green('r')} for re-run Scully, or ${green('q')} for close the servers.
 ${yellow('------------------------------------------------------------')}`
-    : ''
-}
+      : ''}
 `);
-    if (stats) {
-      const scullyStatsFilePath = join(scullyConfig.homeFolder, 'scullyStats.json');
-      const pluginTimings = totalPluginTimes(durations);
-      const scullyStats = {
-        numberOfRoutes,
-        generatingTime: Math.floor(seconds * 100) / 100,
-        routesPerSecond: routesProSecond,
-        findingRoutesAngular: durations.Traverse ? durations.Traverse / 1000 : '',
-        routeDiscovery: durations.Discovery / 1000,
-        renderingPages: durations.Render / 1000,
-        pluginTimings,
-      };
-      Object.entries(pluginTimings).forEach(([name, duration]) =>
-        log(`${name.padEnd(40, ' ')} - ${(Math.floor(duration * 100) / 100).toString().padStart(10, ' ')}`)
-      );
-      writeFileSync(scullyStatsFilePath, JSON.stringify(scullyStats, undefined, 4));
-    }
-    if (!watch) {
-      process.exit(0);
-    }
-  });
+  if (stats) {
+    const scullyStatsFilePath = join(scullyConfig.homeFolder, 'scullyStats.json');
+    const pluginTimings = totalPluginTimes(durations);
+    const scullyStats = {
+      numberOfRoutes,
+      generatingTime: Math.floor(seconds * 100) / 100,
+      routesPerSecond: routesProSecond,
+      findingRoutesAngular: durations.Traverse ? durations.Traverse / 1000 : '',
+      routeDiscovery: durations.Discovery / 1000,
+      renderingPages: durations.Render / 1000,
+      pluginTimings,
+    };
+    Object.entries(pluginTimings).forEach(([name, duration]) => log(`${name.padEnd(40, ' ')} - ${(Math.floor(duration * 100) / 100).toString().padStart(10, ' ')}`)
+    );
+    writeFileSync(scullyStatsFilePath, JSON.stringify(scullyStats, undefined, 4));
+  }
 };
+
 
 function measurePerformance(resolve: (value?: unknown) => void): PerformanceObserverCallback {
   return (list, observer) => {
